feat: [AICPS-55] Change workflow and refactor code in LandingPage
