+++ src/pages/HomePage/Homepage.tsx
@@ -1,19 +1,13 @@
-import { useCallback, useRef } from 'react';
+import { useRef } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 
-import UserPage from '../UserPage/UserPage';
-import BannerScreen from '@app/components/LandingPage/BannerScreen';
-import FeatureSection from '@app/components/LandingPage/FeatureSection';
-import LandingLayout from '@app/components/LandingPage/LandingLayout';
-import LazyComponent from '@app/components/LandingPage/LazyComponent';
-import MainScreen from '@app/components/LandingPage/MainScreen';
-import StepSection from '@app/components/LandingPage/StepSection';
 import FaqSection from '@app/components/Layout/FaqSection/FaqSection';
 import FooterSection from '@app/components/Layout/Footer/FooterSection';
 import BannerUserScreen from '@app/components/UserPage/BannerUserScreen';
 import StepScreen from '@app/components/UserPage/StepScreen';
-import { smoothScrollTo } from '@app/utils/scroll';
+import FeatureSection from '@app/pages/HomePage/LandingPage/FeatureSection';
+import StepSection from '@app/pages/HomePage/LandingPage/StepSection';
 
 const Homepage = () => {
   const isAuth = useSelector((state: any) => state.auth.isAuth);
